+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement Bugzilla metadata collector
 - Implement validation for for affects with exceptional combination of affectedness and resolution (OSIDB-361)
 - Implement validation for services related products with WONTREPORT resolution (OSIDB-362)
+- Implement validation for combinations of affectedness and resolution (OSIDB-360)
 
 ### Changed
 - Change logging of celery and django to filesystem (OSIDB-418)